# Unified allocation throttling