… module/zfs/metaslab.c
… module/zfs/zio.c