@allanjude 2025-02-21
This looks great, I especially like the concept of maintaining the minimum queue depth to keep all of the devices busy.

I know in general you are against adding tunables, but I wonder if a few of the magic numbers could be controllable.
Are there any kstats that we might want to add to be able to measure how this is working under load?