@amotin 2025-02-21
> I know in general you are against adding tunables, but I wonder if a few of the magic numbers could be controllable.

I was thinking about some, but I've decided there is a pretty thin margin between different factors where the algorithm would work as planned, and it would be much easier to mess it up by random tuning rather than improve.  I'll think about some that could have sense.

> Are there any kstats that we might want to add to be able to measure how this is working under load?

I've used some custom tools to collect some data while benchmarking the code, but not sure what I would expose as kstat.  I might think about it, but ideas are welcome.