@amotin 2025-02-24
@allanjude I've added `metaslab_perf_bias` tunable, controlling performance-based metaslab bias between never, on busy pool or always, to cover my above doubts.