@richardm1 2025-03-07
I'm curious about the impact of these changes to _read_ performance; specifically when reading back data _written_ through this new balance mechanism.  Has this aspect been tested?

I mention this because I've been working to optimize sequential read performance of a small system with unbalanced (in both size and performance) rotating mirrors.  I've noticed how the data "lands" on the platters can have a significant perf. impact when reading back said data later on.

On this box the most relevant and impactful knobs are `metaslab_aliquot` (lifting usually helps but only to a point) and `metaslab_bias_enabled` (disabling definitely helps) so a completely new write allocator is relevant to my interests.  Write speed is far less important to me.

Is the code in a usable state where I (as a non-programmer who cannot even spell `git`) can kick the tires on it? I'd be thrilled to help with testing.