@amotin 2025-03-07
@richardm1 The mentioned module parameters are still applicable and working, and aside of me increasing `metaslab_aliquot` here to 2MB, considering slowly increasing linear density of modern HDDs, the on-disk placement should be similar, at least to what was designed originally, but later semi-broken.  I don't have many read benchmarks for this, but you are welcome to try.  The code should be totally usable to the best of my knowledge.  I have no open issues about it.